+++ src/main/lombok/org/heigit/ohsome/ohsomeapi/oshdb/DbConnData.java
@@ -1,6 +1,6 @@
 package org.heigit.ohsome.ohsomeapi.oshdb;
 
-import com.zaxxer.hikari.HikariConfig;
+import javax.sql.DataSource;
 import org.heigit.ohsome.oshdb.api.db.OSHDBDatabase;
 import org.heigit.ohsome.oshdb.api.db.OSHDBJdbc;
 import org.heigit.ohsome.oshdb.util.tagtranslator.TagTranslator;
@@ -11,8 +11,8 @@ public class DbConnData {
   public static OSHDBDatabase db = null;
   public static OSHDBJdbc keytables = null;
   public static TagTranslator tagTranslator = null;
-  public static RemoteTagTranslator mapTagTranslator = null;
-  public static HikariConfig keytablesDbPoolConfig = null;
+  public static DataSource dbSource = null;
+  public static DataSource keytablesDbSource = null;
 
   private DbConnData() {
     throw new IllegalStateException("Utility class");